requirements.txt
asgiref==3.7.2
azure-common==1.1.28
azure-core==1.27.1
azure-storage-blob==12.16.0
azure-storage-common==2.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
Django>=3.2,<4.0
django-appconf==1.0.5
django-compressor==4.3.1
django-crispy-forms==2.0
django-storages==1.13.2
django-taggit==4.0.0
idna==3.4
isodate==0.6.1
Pillow==9.5.0
pycparser==2.21
python-dateutil==2.8.2
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
